backend/pyproject.toml
[project]
name = "qdrant-retrieval-agent"
version = "0.1.0"
description = "Qdrant-based retrieval agent for Physical AI & Humanoid Robotics book content"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.1",
    "qdrant-client[fastembed]>=1.14.2",
    "openai>=1.0.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.24.0",
    "cohere>=5.20.0",
    "rich>=14.2.0",
]
